requirements.txt
flask
plankapy
mysql-connector-python